Move customHook of useOscillator to respective file
